Choose pulse to start ADMA FSM based on transfer direction.
As noted in 5dc30d1, the previous fix to use the cmd_start_puls to move the ADMA FSM out of the IDLE state had a chance of having data to send to the card "too soon." It turns out that in such a situation, the data doesn't go out too soon, but it never goes out and leaves the eMMC device waiting for the write data forever. This is because the start_write pulse sent to the data serializer comes at a point where its FSM is still waiting for the command response. The remedy is to use cmd_start_puls for block reads and cmd_compl_puls for block writes.